requirements.txt
Appium-Python-Client>=0.25
grpcio>=1.11.0
grpcio-tools>=1.11.0
pbr==4.0.2
Pillow>=5.0.0
PyHamcrest==1.9.0
pytest>=3.6.0
pytest-base-url>=1.4.1
pytest-html==1.15.2
pytest-metadata>=1.6.0
pytest-mock==1.6.3
pytest-rerunfailures==1.0.0
pytest-selenium>=1.12.0
pytest-timeout==1.3.0
pytest-variables==1.7.0
pytest-xdist==1.18.1
python-jenkins==0.4.15
pyyaml>=4.2b1
requests>=2.20.0
selenium==3.7.0
flaky>=3.5.0